+++ app/connectors_service/connectors/sources/mssql.py
@@ -25,7 +25,7 @@ from sqlalchemy import create_engine, text
 from sqlalchemy.engine import URL
 from sqlalchemy.exc import ProgrammingError
 
-from connectors.sources.generic_database import (
+from app.connectors_service.connectors.sources.shared.database.generic_database import (
     DEFAULT_FETCH_SIZE,
     DEFAULT_RETRY_COUNT,
     DEFAULT_WAIT_MULTIPLIER,
